workflow: limit & top-n: migrate from u32 -> u64

+++ miso-connectors/src/lib.rs
@@ -200,7 +200,7 @@ pub trait Connector: Debug + Send + Sync {
 
     /// Returns the handle with limit to predicate pushdown.
     /// None means it can't predicate pushdown limit.
-    fn apply_limit(&self, _max: u32, _handle: &dyn QueryHandle) -> Option<Box<dyn QueryHandle>> {
+    fn apply_limit(&self, _max: u64, _handle: &dyn QueryHandle) -> Option<Box<dyn QueryHandle>> {
         None
     }
 
@@ -209,7 +209,7 @@ pub trait Connector: Debug + Send + Sync {
     fn apply_topn(
         &self,
         _sorts: &[Sort],
-        _max: u32,
+        _max: u64,
         _handle: &dyn QueryHandle,
     ) -> Option<Box<dyn QueryHandle>> {
         None
